Move Class definition creation to preimport
Almost all functions that rely on the custom-defined
PowerShell Classes
(EncryptedPassword, PasswordResult and PasswordHistory)
constantly call recreate those classes, which causes overhead.
By moving that creation to the Preimport.ps1,
they are available in the module as soon as it is imported.
Therefore all references in other files can be removed.

Fixes: #83