Updated all PasswordStatePassword functions. Added new class PasswordResetResult (#120)
* Added Alias MoveToRecycleBin (Original API parameter name) to SendToRecycleBin.
Added output if delete request was send successfully.
Changed the boolean variables from uppercase True/False to lowercase true/false (API Documentation:  Boolean fields must be specified in the lowercase format i.e. true and false, not True and False. )

* Fixed/Added Reason (Header not Headers)

* replaced write-verbose with PSFramework

* added preventauditing method

* Changed type of PasswordID, PasswordListID and AccountTypeID from [int] to [Nullable[System.Int32]]. The value of these variables would be 0 before this change if the response from the api does not contain PasswordID, PasswordListID or AccountTypeID. Now it is $null (empty).

* improved exception/error handling and changed passwordlistid parameter to accept $null as value

* added new class PasswordResetResult.
if password reset is enabled for an object and you want to change the password field (Set-PasswordStatePassword) your request will be queued (Queued for Reset).
Then the returned object from the api is different to the normal password object and we need a new class for this.

* added validation for the password field to exclude german umlauts and other latin/non-latin diacritics or invalid characters that the api does not understand.
Removed PreventAuditing since it is not supported in POST/PUT requests.
Added support for new Environment setting for PasswordsInPlainText.
Added error/exception handling.

* Updated Set-PasswordStatePassword to support all available api methods/properties.
Added Reason to the Header.
Rewrite the bound properties part to also support new api properties that are not part of the password result we first get to validate if the password was found.
Added support for newly created PasswordResetResult class to support the Status return about Queued password resets from the api.
Added error/exception handling.

* added reason

* Updated documentation, added Alias MoveToRecycleBin

* added description for reason parameter

* Added documentation for updated Set-PasswordStatePassword function

* Removed the powershell examples in the description and added a description instead (platyPS does not like this). Changed the INPUTS/OUTPUTS.

* Updated external help file

* updated list of exceptions

* Added AccountType to the pester test since AccountType definition now requrires at least 1 character (length 1-50)

* An empty accounttype is also possible, removed validatelength

* updated variable types to an array of nullable integers to workaround the platyps bug

* updated external help file

* updated INPUTS/OUTPUTS

* updated INPUTS/OUTPUTS

* updated INPUTS/OUTPUTS in external help file

* Removed unessecary whitespaces.
Removed Workflow as CommandType in Get-Command due to the error for powershell 7: "Error: "Unable to match the identifier name Workflow to a valid enumerator name. Specify one of the following enumerator names and try again:
Alias, Function, Filter, Cmdlet, ExternalScript, Application, Script, Configuration, All"

* Workaround for Help tests error that parameter mandatory status is wrong: Set Mandatory on all parameter sets for these two parameters to false.